pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "imgproc"
version = "1.0.0"
description = "Intelligent product-image resizer for Delicious Display"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0",
    "numpy>=1.24",
    "pydantic>=2.5",
    "PyYAML>=6.0",
    "click>=8.1",
    "jinja2>=3.1",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "openpyxl>=3.1",
    "imagehash>=4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
imgproc = "imgproc.cli:main"
imgproc-ui = "imgproc.web:run_server"
imgproc-sort = "imgproc.ingest.sort:main"
imgproc-hero = "imgproc.ingest.hero:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
imgproc = ["report/template.html.j2", "web/static/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
